feat(edge): add fdo feature flag
