+++ app/react/portainer/feature-flags/useRedirectFeatureFlag.ts
@@ -1,22 +1,6 @@
 import { useRouter } from '@uirouter/react';
 
-import { usePublicSettings } from '@/react/portainer/settings/queries';
-
-export enum FeatureFlag {}
-
-export function useFeatureFlag(
-  flag: FeatureFlag,
-  {
-    onSuccess,
-    enabled = true,
-  }: { onSuccess?: (isEnabled: boolean) => void; enabled?: boolean } = {}
-) {
-  return usePublicSettings<boolean>({
-    select: (settings) => settings.Features[flag],
-    onSuccess,
-    enabled,
-  });
-}
+import { FeatureFlag, useFeatureFlag } from './useFeatureFlag';
 
 export function useRedirectFeatureFlag(
   flag: FeatureFlag,
